Organise files into structured packages
