+++ src/main/java/friday/task/Deadline.java
@@ -1,4 +1,4 @@
-package friday;
+package friday.task;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
