+++ src/main/java/friday/task/Event.java
@@ -1,4 +1,4 @@
-package friday;
+package friday.task;
 
 /**
  * Represents an event task in the Duke application.
