+++ src/main/java/friday/task/Task.java
@@ -1,4 +1,6 @@
-package friday;
+package friday.task;
+
+import friday.Item;
 
 /**
  * Represents a task in the Friday application.
