+++ src/main/java/friday/util/Storage.java
@@ -1,4 +1,4 @@
-package friday;
+package friday.util;
 
 import java.io.File;
 import java.io.FileWriter;
